Add files via upload
This is our first work through on the new pgls function called "pgls.sey". The ".sey" attached to the end of pgls stands for the standard error of Y (meaning your response variable in a regression. As of now (12/01/21) we are calculating the standard error of a species as sqrt(1/TotalIndividuals). The pgls.sey allows us to take the standard error into account, practically weighing the regression by the sample size of each species, as opposed to the regular pgls function which has no weigh options